+++ src/mantle/view/engines/class-engine-resolver.php
@@ -18,7 +18,7 @@ class Engine_Resolver {
 	/**
 	 * The array of engine resolvers.
 	 *
-	 * @var array<string, \Closure>
+	 * @var array<string, \Closure(): Engine>
 	 */
 	protected array $resolvers = [];
 
@@ -34,30 +34,29 @@ class Engine_Resolver {
 	 *
 	 * The engine string typically corresponds to a file extension.
 	 *
-	 * @param  string   $engine
-	 * @param  \Closure $resolver
+	 * @param  string             $engine
+	 * @param  \Closure(): Engine $resolver
 	 */
-	public function register( $engine, Closure $resolver ): void {
-			unset( $this->resolved[ $engine ] );
+	public function register( string $engine, Closure $resolver ): void {
+		unset( $this->resolved[ $engine ] );
 
-			$this->resolvers[ $engine ] = $resolver;
+		$this->resolvers[ $engine ] = $resolver;
 	}
 
 	/**
 	 * Resolve an engine instance by name.
 	 *
-	 * @param  string $engine
-	 * @return Engine|\Illuminate\View\Engines\CompilerEngine
+	 * @param string $engine
 	 *
 	 * @throws InvalidArgumentException Thrown on unknown engine.
 	 */
-	public function resolve( $engine ) {
+	public function resolve( string $engine ): Engine {
 		if ( isset( $this->resolved[ $engine ] ) ) {
 			return $this->resolved[ $engine ];
 		}
 
 		if ( isset( $this->resolvers[ $engine ] ) ) {
-			$this->resolved[ $engine ] = call_user_func( $this->resolvers[ $engine ] );
+			$this->resolved[ $engine ] = $this->resolvers[ $engine ]();
 			return $this->resolved[ $engine ];
 		}