+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `Blade` facade with `Blade::render_string()` method to render a string
+  using the Blade templating engine.
 - Added `assertBlockExists()` and `assertBlockMissing()` to the response assertions.
 - Added `$count` param to `assertPostExists()`.
 - Added dynamic sentence length support to `paragraph()` and `paragraphs()` on
   the Faker block provider.
 - Added `Environment` attribute to allow for environment-specific code to be run during tests.
 - Added support for user agent setting in tests via the `#[User_Agent]` attribute.
+- Added `slash()` helper to factories to automatically slash content before storing.
 
 ### Changed